requirements.txt
pandas>=2.0.0
numpy>=1.24.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
boto3>=1.34.0
botocore[crt]>=1.34.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
rapidfuzz>=3.5.0
ruff>=0.1.0
lxml>=6.0.2
